requirement.txt
astroid==2.3.0
boto3==1.9.239
botocore==1.12.239
colorama==0.4.1
Django==2.2.5
django-crispy-forms==1.7.2
django-storages==1.7.2
docutils==0.15.2
gunicorn==19.9.0
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==6.1.0
pylint==2.4.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
